add pool property isProcessingConfirmedClaims

+++ contracts/instance/module/IPolicy.sol
@@ -66,8 +66,8 @@ interface IPolicy {
     struct ClaimInfo {
         Amount claimAmount;
         Amount paidAmount;
-        uint8 payoutsCount;
-        uint8 openPayoutsCount;
+        uint24 payoutsCount;
+        uint24 openPayoutsCount;
         bytes submissionData; // use case specific claim submission data, no changes after submitting the claim
         bytes processData; // use case specific data that may include information supporting confirm or decline
         Timestamp closedAt; // payment of confirmed claim amount (or declinedAt)
